Create testInitializer.go, containing test utilities
