--- mock/mockVMOutput.go
@@ -1,54 +1,0 @@
-package mock
-
-import (
-	"math/big"
-
-	"github.com/ElrondNetwork/elrond-vm-common"
-)
-
-func MakeVMOutput() *vmcommon.VMOutput {
-	return &vmcommon.VMOutput{
-		ReturnCode:      vmcommon.Ok,
-		ReturnMessage:   "",
-		ReturnData:      make([][]byte, 0),
-		GasRemaining:    0,
-		GasRefund:       big.NewInt(0),
-		DeletedAccounts: make([][]byte, 0),
-		TouchedAccounts: make([][]byte, 0),
-		Logs:            make([]*vmcommon.LogEntry, 0),
-		OutputAccounts:  make(map[string]*vmcommon.OutputAccount),
-	}
-}
-
-func AddFinishData(vmOutput *vmcommon.VMOutput, data []byte) {
-	vmOutput.ReturnData = append(vmOutput.ReturnData, data)
-}
-
-func AddNewOutputAccount(vmOutput *vmcommon.VMOutput, address []byte, balanceDelta int64, data []byte) *vmcommon.OutputAccount {
-	account := &vmcommon.OutputAccount{
-		Address:        address,
-		Nonce:          0,
-		BalanceDelta:   big.NewInt(balanceDelta),
-		Balance:        big.NewInt(0),
-		StorageUpdates: make(map[string]*vmcommon.StorageUpdate),
-		Code:           nil,
-		Data:           data,
-	}
-	vmOutput.OutputAccounts[string(address)] = account
-	return account
-}
-
-func SetStorageUpdate(account *vmcommon.OutputAccount, key []byte, data []byte) {
-	keyString := string(key)
-	update, exists := account.StorageUpdates[keyString]
-	if !exists {
-		update = &vmcommon.StorageUpdate{}
-		account.StorageUpdates[keyString] = update
-	}
-	update.Offset = key
-	update.Data = data
-}
-
-func SetStorageUpdateStrings(account *vmcommon.OutputAccount, key string, data string) {
-	SetStorageUpdate(account, []byte(key), []byte(data))
-}
